Alternative persistence support alpha

+++ packages/store/src/persistence/PersistenceRebaser.ts
@@ -7,13 +7,22 @@ import {
 } from '@verdant-web/common';
 import { Context } from '../context/context.js';
 import { AbstractTransaction, PersistenceMetadataDb } from './interfaces.js';
+import type { PersistenceMetadata } from './PersistenceMetadata.js';
 
 export class PersistenceRebaser {
 	constructor(
 		private db: PersistenceMetadataDb,
+		private meta: PersistenceMetadata,
 		private ctx: Pick<
 			Context,
-			'closing' | 'log' | 'time' | 'internalEvents' | 'globalEvents' | 'config'
+			| 'closing'
+			| 'log'
+			| 'time'
+			| 'internalEvents'
+			| 'globalEvents'
+			| 'config'
+			| 'closeLock'
+			| 'persistenceShutdownHandler'
 		>,
 	) {}
 
@@ -22,9 +31,12 @@ export class PersistenceRebaser {
 	 * keep storage clean for non-syncing clients by compressing history.
 	 */
 	tryAutonomousRebase = async () => {
-		const localReplicaInfo = await this.db.getLocalReplica();
+		const localReplicaInfo = await this.meta.getLocalReplica();
 		if (localReplicaInfo.lastSyncedLogicalTime) return; // cannot autonomously rebase if we've synced
+		if (this.ctx.closing || this.ctx.persistenceShutdownHandler.isShuttingDown)
+			return;
 		// but if we have never synced... we can rebase everything!
+		this.ctx.log('debug', 'Running autonomous library rebase');
 		await this.runRebase(this.ctx.time.now);
 	};
 
@@ -35,47 +47,55 @@ export class PersistenceRebaser {
 	 * their undo stack.
 	 */
 	private runRebase = async (globalAckTimestamp: string) => {
-		if (this.ctx.closing) return;
+		if (this.ctx.closing || this.ctx.persistenceShutdownHandler.isShuttingDown)
+			return;
 
-		// find all operations before the global ack
-		let lastTimestamp;
-		const toRebase = new Set<ObjectIdentifier>();
-		const transaction = this.db.transaction({
-			storeNames: ['baselines', 'operations'],
-			mode: 'readwrite',
-		});
-		let operationCount = 0;
-		await this.db.iterateAllOperations(
-			(patch) => {
-				toRebase.add(patch.oid);
-				lastTimestamp = patch.timestamp;
-				operationCount++;
-			},
+		await this.db.transaction(
 			{
-				before: globalAckTimestamp,
-				transaction,
+				storeNames: ['baselines', 'operations'],
+				mode: 'readwrite',
+			},
+			async (transaction) => {
+				// find all operations before the global ack
+				const toRebase = new Set<ObjectIdentifier>();
+				let lastTimestamp;
+				let operationCount = 0;
+				await this.db.iterateAllOperations(
+					(patch) => {
+						toRebase.add(patch.oid);
+						lastTimestamp = patch.timestamp;
+						operationCount++;
+					},
+					{
+						before: globalAckTimestamp,
+						transaction,
+					},
+				);
+
+				if (!toRebase.size) {
+					return;
+				}
+
+				if (
+					this.ctx.closing ||
+					this.ctx.persistenceShutdownHandler.isShuttingDown
+				) {
+					return;
+				}
+
+				// rebase each affected document
+				let newBaselines = [];
+				for (const oid of toRebase) {
+					newBaselines.push(
+						await this.rebase(
+							oid,
+							lastTimestamp || globalAckTimestamp,
+							transaction,
+						),
+					);
+				}
 			},
 		);
-
-		if (!toRebase.size) {
-			return;
-		}
-
-		if (this.ctx.closing) {
-			return;
-		}
-
-		// rebase each affected document
-		let newBaselines = [];
-		for (const oid of toRebase) {
-			newBaselines.push(
-				await this.rebase(
-					oid,
-					lastTimestamp || globalAckTimestamp,
-					transaction,
-				),
-			);
-		}
 		this.ctx.globalEvents.emit('rebase');
 	};
 
@@ -92,26 +112,28 @@ export class PersistenceRebaser {
 			this.ctx.config.persistence?.rebaseTimeout ?? 10000,
 			timestamp,
 		);
+		this.ctx.log('debug', 'Scheduled rebase up to global ack', timestamp);
 	};
 	private rebaseTimeout: NodeJS.Timeout | null = null;
 
-	rebase = async (
+	private rebase = async (
 		oid: ObjectIdentifier,
 		upTo: string,
-		providedTx?: AbstractTransaction,
+		transaction: AbstractTransaction,
 	) => {
-		const transaction =
-			providedTx ||
-			this.db.transaction({
-				storeNames: ['operations', 'baselines'],
-				mode: 'readwrite',
-			});
+		if (this.ctx.closing || this.ctx.persistenceShutdownHandler.isShuttingDown)
+			return;
+
 		const baseline = await this.db.getBaseline(oid, { transaction });
 		let current: any = baseline?.snapshot || undefined;
 		let operationsApplied = 0;
 		let authz = baseline?.authz;
 		const deletedRefs: Ref[] = [];
-		await this.db.consumeEntityOperations(
+
+		if (this.ctx.closing || this.ctx.persistenceShutdownHandler.isShuttingDown)
+			return;
+
+		await this.db.iterateEntityOperations(
 			oid,
 			(patch) => {
 				// FIXME: this seems like the wrong place to do this
@@ -139,32 +161,45 @@ export class PersistenceRebaser {
 			timestamp: upTo,
 			authz,
 		};
-		if (newBaseline.snapshot) {
-			await this.db.setBaselines([newBaseline], { transaction });
-		} else {
-			await this.db.deleteBaseline(oid, { transaction });
-		}
 
-		this.ctx.log(
-			'debug',
-			'rebased',
-			oid,
-			'up to',
-			upTo,
-			':',
-			current,
-			'and deleted',
-			operationsApplied,
-			'operations',
-		);
+		// still time to cancel now...
+		if (this.ctx.closing || this.ctx.persistenceShutdownHandler.isShuttingDown)
+			return;
 
-		// cleanup deleted refs
-		if (deletedRefs.length) {
-			const fileRefs = deletedRefs.filter(isFileRef);
-			if (fileRefs.length) {
-				this.ctx.internalEvents.emit('filesDeleted', fileRefs);
+		// FROM HERE, WE ARE COMMITTED TO THE REBASE -- otherwise data will be corrupted.
+		this.ctx.closeLock = (async () => {
+			if (newBaseline.snapshot) {
+				await this.db.setBaselines([newBaseline], { transaction });
+			} else {
+				await this.db.deleteBaseline(oid, { transaction });
 			}
-		}
+
+			await this.db.deleteEntityOperations(oid, {
+				to: upTo,
+				transaction,
+			});
+
+			this.ctx.log(
+				'debug',
+				'rebased',
+				oid,
+				'up to',
+				upTo,
+				':',
+				current,
+				'and deleted',
+				operationsApplied,
+				'operations',
+			);
+
+			// cleanup deleted refs
+			if (deletedRefs.length) {
+				const fileRefs = deletedRefs.filter(isFileRef);
+				if (fileRefs.length) {
+					this.ctx.internalEvents.emit('filesDeleted', fileRefs);
+				}
+			}
+		})();
 
 		return newBaseline;
 	};
